format fix

+++ app/src/main/java/com/wacom/will3/ink/raster/rendering/demo/MainActivity.kt
@@ -31,6 +31,7 @@ import com.wacom.will3.ink.raster.rendering.demo.serialization.InkEnvironmentMod
 import com.wacom.will3.ink.raster.rendering.demo.tools.raster.*
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.newSingleThreadContext
 import top.defaults.colorpicker.ColorPickerPopup
@@ -38,6 +39,7 @@ import top.defaults.colorpicker.ColorPickerPopup.ColorPickerObserver
 import java.io.File
 import java.io.FileInputStream
 import java.io.FileOutputStream
+import java.util.*
 
 class MainActivity : AppCompatActivity(), RasterView.InkingSurfaceListener {
 
@@ -58,45 +60,42 @@ class MainActivity : AppCompatActivity(), RasterView.InkingSurfaceListener {
     private var defaultDrawingTool = PencilTool.uri
     private var drawingTool: RasterTool? = null
 
-    private var popupWindow: PopupWindow? = null
+    private lateinit var popupWindow: PopupWindow
 
     private var drawingColor: Int = Color.argb(255, 74, 74, 74)
     private var lastEvent: MotionEvent? = null
 
-    private var tempStrokeFile: String? = null
-    private var refreshing: Boolean = false
-    private var currentBackground: Int = 3
+    private var tempStrokeFile = ""
+    private var refreshing = false
+    private var currentBackground = 3
 
-    private val scope = CoroutineScope(newSingleThreadContext("synchronizationPool"))
+    fun add(view:View){
+        //if(rasterDrawingSurface.layerPos<2)rasterDrawingSurface.layerPos++
+    }
+
+    fun minus(view:View){
+        //if(rasterDrawingSurface.layerPos>0)rasterDrawingSurface.layerPos--
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        // We maintain he same orientation
-        //disableRotation()
-
         if (savedInstanceState != null) {
             drawingColor = savedInstanceState.getInt("selectedColor")
-            if (savedInstanceState.getString("selectedTool") != null) {
-                defaultDrawingTool = savedInstanceState.getString("selectedTool")!!
-            }
-            tempStrokeFile = savedInstanceState.getString("strokes")
+            defaultDrawingTool = savedInstanceState.getString("selectedTool") ?: ""
+            tempStrokeFile = savedInstanceState.getString("strokes") ?: ""
             currentBackground = savedInstanceState.getInt("currentBackground")
             refreshing = true
-
         } else {
             resetInkModel()
         }
 
-        inkEnvironmentModel =
-            InkEnvironmentModel(this) // Initializes the environment data for serialization
+        inkEnvironmentModel = InkEnvironmentModel(this) // Initializes the environment data for serialization
 
         setColor(drawingColor) //set default color
 
-        background_waiting.setOnTouchListener { _, event ->
-            true
-        }
+        background_waiting.setOnTouchListener { _, event -> true }
 
         rasterDrawingSurface.setOnTouchListener { _, event ->
             // We save the last event just in case we receive a CANCEL action
@@ -105,19 +104,12 @@ class MainActivity : AppCompatActivity(), RasterView.InkingSurfaceListener {
             if ((event.action == MotionEvent.ACTION_DOWN) ||
                 (event.action == MotionEvent.ACTION_MOVE) ||
                 (event.action == MotionEvent.ACTION_UP)
-            ) {
-                lastEvent = MotionEvent.obtain(event)
-            } else {
-                lastEvent?.action = MotionEvent.ACTION_UP //we convert the latest event in END event
-            }
+            ) lastEvent = MotionEvent.obtain(event)
+            else lastEvent?.action = MotionEvent.ACTION_UP //we convert the latest event in END event
 
-            if (lastEvent != null) {
-                rasterDrawingSurface.surfaceTouch(lastEvent!!)
-            }
+            if (lastEvent != null) rasterDrawingSurface.surfaceTouch(lastEvent!!)
 
-            if (event.action == MotionEvent.ACTION_UP) {
-                lastEvent = null
-            }
+            if (event.action == MotionEvent.ACTION_UP) lastEvent = null
             true
         }
 
@@ -133,15 +125,11 @@ class MainActivity : AppCompatActivity(), RasterView.InkingSurfaceListener {
         // save the current selected tool, color and strokes
         refreshing = true
         val tempFile = File(cacheDir, "tempStrokes.uim")
-        if (tempFile.exists()) {
-            tempFile.delete()
-        }
+        if (tempFile.exists()) tempFile.delete()
         save(Uri.fromFile(tempFile))
-
         savedInstanceState.putString("strokes", tempFile.absolutePath)
         savedInstanceState.putString("selectedTool", drawingTool?.uri())
         savedInstanceState.putInt("selectedColor", drawingColor)
-
         savedInstanceState.putInt("currentBackground", currentBackground)
     }
 
@@ -151,13 +139,8 @@ class MainActivity : AppCompatActivity(), RasterView.InkingSurfaceListener {
         // Make sure the request was successful
         if (resultCode == Activity.RESULT_OK) {
             // Check which request we're responding to
-            if (requestCode == OPEN_FILE_ACTION) {
-                scope.launch {
-                    load(data!!.data!!)
-                }
-            } else if (requestCode == CREATE_FILE_ACTION) {
-                save(data!!.data!!)
-            }
+            if (requestCode == OPEN_FILE_ACTION) CoroutineScope(Dispatchers.Default).launch { load(data!!.data!!) }
+            else if (requestCode == CREATE_FILE_ACTION) save(data!!.data!!)
         }
     }
 
@@ -207,50 +190,45 @@ class MainActivity : AppCompatActivity(), RasterView.InkingSurfaceListener {
 
     fun setTool(view: View, tool: RasterTool) {
         drawingTool = tool
-
         val dt = drawingTool as RasterTool
         rasterDrawingSurface.setTool(dt)
         highlightTool(view)
     }
 
     fun highlightTool(view: View) {
-        btn_pencil.setActivated(false)
-        btn_water_brush.setActivated(false)
-        btn_ink_brush.setActivated(false)
-        btn_crayon.setActivated(false)
-        btn_eraser.setActivated(false)
-        view.setActivated(true)
+        btn_pencil.isActivated = false
+        btn_water_brush.isActivated = false
+        btn_ink_brush.isActivated = false
+        btn_crayon.isActivated = false
+        btn_eraser.isActivated = false
+        view.isActivated = true
     }
 
     fun load(view: View) {
         val intent = Intent(Intent.ACTION_OPEN_DOCUMENT)
         intent.addCategory(Intent.CATEGORY_OPENABLE)
-        intent.setType("*/*");
+        intent.type = "*/*";
         startActivityForResult(intent, OPEN_FILE_ACTION)
     }
 
     fun load(path: Uri) {
-        this@MainActivity.runOnUiThread(java.lang.Runnable {
-            if (!refreshing) {
-                Toast.makeText(this, "Loading..", Toast.LENGTH_SHORT).show()
-            }
-
+        this@MainActivity.runOnUiThread{
+            if (!refreshing) Toast.makeText(this, "Loading..", Toast.LENGTH_SHORT).show()
             background_waiting.visibility = View.VISIBLE
-
             // Before loading we clear the screen
             rasterDrawingSurface.clear();
 
-        })
+        }
 
         try {
-            getContentResolver().openFileDescriptor(path, "r").use { pfd ->
+            contentResolver.openFileDescriptor(path, "r").use { pfd ->
                 FileInputStream(pfd?.fileDescriptor).use { fileInputStream ->
 
                     val bytes = fileInputStream.readBytes()
                     inkModel = Will3Codec.decode(bytes)
-                    mainGroup = inkModel.inkTree!!.root!! as StrokeGroupNode
+                    mainGroup = inkModel.inkTree.root!! as StrokeGroupNode
 
-                    val it = inkModel.inkTree!!.root!!.iterator()
+                    val it = inkModel.inkTree.root!!.iterator()
                     while (it.hasNext()) {
                         val stroke = it.next()
                         if (stroke is StrokeNode) {
@@ -266,71 +244,40 @@ class MainActivity : AppCompatActivity(), RasterView.InkingSurfaceListener {
                                     var sensorChannelList: List<SensorChannel>? = null
                                     val sensorDataID = stroke.data.sensorDataID
                                     if (sensorDataID != null) {
-                                        rasterDrawingSurface.addSensorData(
-                                            inkModel.sensorDataRepository.get(
-                                                sensorDataID
-                                            )!!
-                                        )
-
-                                        val inputContextId =
-                                            inkModel.sensorDataRepository.get(sensorDataID)?.inputContextId
+                                        rasterDrawingSurface.addSensorData(inkModel.sensorDataRepository.get(sensorDataID)!!)
+                                        val inputContextId = inkModel.sensorDataRepository.get(sensorDataID)?.inputContextId
                                         if (inputContextId != null) {
-                                            val sensorContextId =
-                                                inkModel.inputConfiguration.getInputContext(
-                                                    inputContextId
-                                                )
-                                                    ?.sensorContextId
-                                            if (sensorContextId != null) {
-                                                sensorChannelList =
-                                                    inkModel.inputConfiguration.getSensorContext(
-                                                        sensorContextId
-                                                    )?.getSensorChannelsContexts()?.first()
-                                                        ?.getAll()
-                                            }
+                                            val sensorContextId = inkModel.inputConfiguration.getInputContext(inputContextId)?.sensorContextId
+                                            if (sensorContextId != null) sensorChannelList = inkModel.inputConfiguration.getSensorContext(sensorContextId)?.getSensorChannelsContexts()?.first()?.getAll()
                                         }
                                     }
-
-                                    this@MainActivity.runOnUiThread(java.lang.Runnable {
-                                        rasterDrawingSurface.drawStroke(
-                                            stroke,
-                                            brush,
-                                            sensorChannelList
-                                        )
-                                    })
-
+                                    this@MainActivity.runOnUiThread{rasterDrawingSurface.drawStroke(stroke,brush,sensorChannelList) }
                                     rasterDrawingSurface.addStroke(stroke, brush)
                                 }
                             }
                         }
                     }
-
                 }
             }
-        } catch (e: Exception) {
-            e.printStackTrace()
+        } catch (e: Exception) { e.printStackTrace()
         }
 
-        this@MainActivity.runOnUiThread(java.lang.Runnable {
-            if (!refreshing) {
-                Toast.makeText(this, "Loading completed.", Toast.LENGTH_SHORT).show()
-            }
+        this@MainActivity.runOnUiThread {
+            if (!refreshing) Toast.makeText(this, "Loading completed.", Toast.LENGTH_SHORT).show()
             background_waiting.visibility = View.GONE
             rasterDrawingSurface.setTool(drawingTool!!)
             setColor(drawingColor)
-        })
+        }
 
         refreshing = false
     }
 
     private fun resetInkModel() {
         inkModel = InkModel()
-
         val root = StrokeGroupNode(Identifier())
-
         inkModel.inkTree.root = root
         mainGroup = StrokeGroupNode(Identifier())
         root.add(mainGroup)
-
         //TODO inkModel.brushRepository.addVectorBrush(brush)
         //TODO splines.clear()
         //TODO paths.clear()
@@ -338,17 +285,15 @@ class MainActivity : AppCompatActivity(), RasterView.InkingSurfaceListener {
 
 
     fun save(uri: Uri) {
-        if (!refreshing) {
-            Toast.makeText(this, "Saving..", Toast.LENGTH_SHORT).show()
-        }
+        if (!refreshing) Toast.makeText(this, "Saving..", Toast.LENGTH_SHORT).show()
 
         inkEnvironmentModel.registerInModel(inkModel)
         inkModel.knowledgeGraph.add(
-            inkModel.inkTree!!.root!!.id.toUUIDString(),
+            inkModel.inkTree.root!!.id.toUUIDString(),
             "created",
             "" + System.currentTimeMillis()
         )
-        inkModel.knowledgeGraph.add(inkModel.inkTree!!.root!!.id.toUUIDString(), "author", "WILL 3")
+        inkModel.knowledgeGraph.add(inkModel.inkTree.root!!.id.toUUIDString(), "author", "WILL 3")
 
 
         // For raster serialization
@@ -357,34 +302,23 @@ class MainActivity : AppCompatActivity(), RasterView.InkingSurfaceListener {
         }
 
         for (stroke in rasterDrawingSurface.strokeNodeList) {
-            if (inkModel.brushRepository.getBrush(stroke.second.name) == null) {
-                inkModel.brushRepository.addRasterBrush(stroke.second)
-            }
-            if (!mainGroup.contains(stroke.first)) {
-                mainGroup.add(stroke.first)
-            }
+            if (inkModel.brushRepository.getBrush(stroke.second.name) == null) inkModel.brushRepository.addRasterBrush(stroke.second)
+            if (!mainGroup.contains(stroke.first)) mainGroup.add(stroke.first)
         }
 
-        getContentResolver().openFileDescriptor(uri, "w").use { pfd ->
-            FileOutputStream(pfd?.fileDescriptor).use { fileOutputStream ->
-                Will3Codec.encode(inkModel, fileOutputStream)
-            }
-        }
+        val fileDescriptor = contentResolver.openFileDescriptor(uri, "w")?.fileDescriptor
+        val outputStream = FileOutputStream(fileDescriptor)
+        Will3Codec.encode(inkModel, outputStream)
     }
 
     override fun onSurfaceCreated() {
-        if (drawingTool != null) {
-            rasterDrawingSurface.setTool(drawingTool!!)
-        } else {
-            selectTool(defaultDrawingTool) // set default tool
-        }
+        if (drawingTool != null) rasterDrawingSurface.setTool(drawingTool!!)
+        else selectTool(defaultDrawingTool) // set default tool
 
-        if (tempStrokeFile != null) {
-            scope.launch {
-                load(Uri.fromFile(File(tempStrokeFile)))
-                File(tempStrokeFile).delete()
-                tempStrokeFile = null
-            }
+        CoroutineScope(Dispatchers.Default).launch {
+            load(Uri.fromFile(File(tempStrokeFile)))
+            File(tempStrokeFile).delete()
+            tempStrokeFile = ""
         }
     }
 
@@ -407,23 +341,15 @@ class MainActivity : AppCompatActivity(), RasterView.InkingSurfaceListener {
         // which view you pass in doesn't matter, it is only used for the window tolken
         val screenPos = IntArray(2)
         view.getLocationOnScreen(screenPos)
-        popupWindow!!.showAtLocation(
-            view,
-            Gravity.NO_GRAVITY,
-            screenPos[0],
-            screenPos[1] + navbar_container.height
-        )
+        popupWindow.showAtLocation(view, Gravity.NO_GRAVITY,screenPos[0],screenPos[1]+navbar_container.height)
     }
 
     fun selectPaper(pos: Int) {
-        if (pos == 0) {
-            changeBackground(R.drawable.btn_paper_01, R.drawable.background1)
-        } else if (pos == 1) {
-            changeBackground(R.drawable.btn_paper_02, R.drawable.background2)
-        } else if (pos == 2) {
-            changeBackground(R.drawable.btn_paper_03, R.drawable.btn_paper_03)
-        } else {
-            changeBackground(R.drawable.btn_paper_04, R.drawable.btn_paper_04)
+        when (pos) {
+            0 -> changeBackground(R.drawable.btn_paper_01, R.drawable.background1)
+            1 -> changeBackground(R.drawable.btn_paper_02, R.drawable.background2)
+            2 -> changeBackground(R.drawable.btn_paper_03, R.drawable.btn_paper_03)
+            else -> changeBackground(R.drawable.btn_paper_04, R.drawable.btn_paper_04)
         }
     }
 
@@ -451,17 +377,15 @@ class MainActivity : AppCompatActivity(), RasterView.InkingSurfaceListener {
     fun changeBackground(background: Int, paper: Int) {
         btnBackground.setImageResource(background)
         drawingLayout.setBackgroundResource(paper)
-        if (popupWindow != null) {
-            popupWindow!!.dismiss()
-        }
+        if(this::popupWindow.isInitialized)popupWindow.dismiss()
     }
 
     fun selectSaveFile(view: View) {
         val intent = Intent(Intent.ACTION_CREATE_DOCUMENT)
-        intent.setType("*/*")
-        intent.putExtra(Intent.EXTRA_TITLE, "ink.uim");
+        intent.type = "*/*"
+        val time = Date().time%10000
+        intent.putExtra(Intent.EXTRA_TITLE, "$time.uim");
         intent.addCategory(Intent.CATEGORY_OPENABLE)
         startActivityForResult(intent, CREATE_FILE_ACTION)
     }
-
 }
